agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.1
galileo==1.39.0
galileo-core==3.79.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.81
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.2.70
langsmith==0.4.37
openai==2.14.0
orjson==3.11.5
packaging==25.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.11.0
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.45
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvloop==0.21.0
wrapt==1.17.3
zstandard==0.25.0
